requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
anyio==4.4.0
arrow==1.3.0
asgiref==3.8.1
attrs==23.2.0
backoff==2.2.1
black==24.4.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses==0.6
datasets==2.20.0
dill==0.3.8
Django==4.2
django-cors-headers==4.4.0
django-filter==24.2
djangorestframework==3.15.2
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
gql==3.5.0
graphql-core==3.2.3
halo==0.0.31
huggingface-hub==0.23.4
idna==3.7
iniconfig==2.0.0
joblib==1.4.2
log-symbols==0.0.14
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
numpy==1.26.4
packaging==24.1
pandas==1.4.4
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
Pygments==2.18.0
pytest==8.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
spinners==0.0.24
sqlparse==0.5.1
tabulate==0.9.0
termcolor==2.4.0
threadpoolctl==3.5.0
tokenizers==0.19.1
tqdm==4.66.4
transform==1.2.0
transform_tools==1.0.5
transformers==4.42.4
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
update-checker==0.18.0
urllib3==2.2.2
validators==0.33.0
xxhash==3.4.1
yarl==1.9.4
